Check in Scintilla version 321
(integrated from komodo (ericp-scintilla-321 branch) change 73425 by ericp)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@10839 b87d16e7-d29f-4ee1-b5ec-817d645d335f

+++ contrib/scintilla/lexers/LexModula.cxx
@@ -92,7 +92,7 @@ static inline bool checkStatement(
 	Accessor &styler,
 	int &curPos,
 	const char *stt, bool spaceAfter = true ) {
-	int len = strlen( stt );
+	int len = static_cast<int>(strlen( stt ));
 	int i;
 	for( i = 0; i < len; i++ ) {
 		if( styler.SafeGetCharAt( curPos + i ) != stt[i] ) {
@@ -113,7 +113,7 @@ static inline bool checkEndSemicolon(
 	int &curPos, int endPos )
 {
 	const char *stt = "END";
-	int len = strlen( stt );
+	int len = static_cast<int>(strlen( stt ));
 	int i;
 	for( i = 0; i < len; i++ ) {
 		if( styler.SafeGetCharAt( curPos + i ) != stt[i] ) {
